+++ R/get_nomalized_counts.R
@@ -32,7 +32,7 @@ get_nomalized_counts <- function(longest_trans_file = NULL,
   # ============================================================================
   # loop read file
   # gene annotation
-  ganao <- read.delim(longest_trans_file,header = F)
+  ganao <- readr::read.delim(longest_trans_file,header = F)
   colnames(ganao) <- c("id","gene_name","gene_id","trans_id","chrom","strand",
                        "cds_rg","exon_rg","5UTR_length","CDS_length","3UTR_length")