+++ NEWS.md
@@ -1,3 +1,9 @@
-# RiboProfiler 0.1.0
+# RiboProfiler update news
 
-* Initial CRAN submission.
+---
+
+## RiboProfiler version 0.1.0 update
+
+- Add **get_track_df** for gene track plot.
+- Add **metagene_plot** for metagene analysis.
+- Add **triAmino_motif_plot** for enriched motif logo plot.